+++ src/features/Post/components/ArticleMenu/components/Menu/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as DropdownMenu from '@radix-ui/react-dropdown-menu'
 import { Article } from 'phosphor-react'
 import { HiddenItems } from './components'
