+++ src/features/Post/components/ArticleMenu/components/index.tsx
@@ -1,1 +1,3 @@
 export * from './Menu'
+
+export * from './ScrollToTop'
