+++ src/features/Post/components/ReadingBarIndicator/index.tsx
@@ -1,21 +1,9 @@
 'use client'
 
-import { useEffect, useState } from 'react'
+import { useScrollBarPercentage } from '@/hooks/useScrollBarPercentage'
 
 export function ReadingBarIndicator() {
-  const [value, setValue] = useState(0)
-
-  useEffect(() => {
-    function handleScroll() {
-      const { scrollTop, scrollHeight } = document.documentElement
-
-      const scrollPercent = (scrollTop / (scrollHeight - window.innerHeight)) * 100
-
-      setValue(scrollPercent)
-    }
-
-    window.addEventListener('scroll', handleScroll)
-  }, [])
+  const value = useScrollBarPercentage()
 
   return (
     <div className={'sticky top-[70px] z-50 h-1 rounded-lg bg-primary opacity-50'} style={{ width: `${value}%` }} />
